requirements.txt
apiclient==1.0.4
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dateutils==0.6.8
Flask==1.1.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml
oauth2client==4.1.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
Pastebin==1.1.2